+++ angular/projects/admin/src/app/admin/document-type/document-type-dialog-create/document-type-dialog-create.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { DocumentTypeService } from '../../../services/document-type.service';
 import { MatDialogRef } from '@angular/material';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'tanam-document-type-dialog-create',
@@ -17,11 +18,22 @@ export class DocumentTypeDialogCreateComponent {
     private readonly formBuilder: FormBuilder,
     private readonly documentTypeService: DocumentTypeService,
     private dialogRef: MatDialogRef<DocumentTypeDialogCreateComponent>,
+    private router: Router
   ) { }
 
   createNewType() {
-    const formData = this.createTypeForm.value;
-    this.documentTypeService.createWithTitle(formData.title);
+    const title = this.createTypeForm.value.title;
+    this.documentTypeService.createWithTitle(this._slugify(title), title);
     this.dialogRef.close();
+    this.router.navigateByUrl(`/_/admin/type/${this._slugify(title)}/edit`);
+  }
+
+  private _slugify(text: string) {
+    return text.toLowerCase()
+      .replace(/\s+/g, '-')
+      .replace(/[^\w\-]+/g, '')
+      .replace(/\-\-+/g, '-')
+      .replace(/^-+/, '')
+      .replace(/-+$/, '');
   }
 }
